==> Notebook_0_Data_Fetching.ipynb <==
# Notebook 0 - Fetching Data and Data Cleaning

# Fetch Data

import pickle
from sklearn.externals import joblib

import pandas as pd

## UCI Madelon

uci_train = 'https://archive.ics.uci.edu/ml/machine-learning-databases/madelon/MADELON/madelon_train.data'

uci_target_url = 'https://archive.ics.uci.edu/ml/machine-learning-databases/madelon/MADELON/madelon_train.labels'

uci_target = pd.read_csv(uci_target_url, header = None) # assigning the target data to 'uci_target'

uci_target.isnull().sum() # checking for nulls; no nulls

uci_train = pd.read_csv(uci_train, delimiter=' ', header=None)

uci_train.isnull().sum().sum() # checking for nulls; nulls are all on column '500'

uci_train = uci_train.drop(500, axis = 1) # dropping the '500' row and reassigning 'uci_train'

uci_train.isnull().sum().sum() # checking that nulls have been dropped and no more nulls exist

uci_train.shape

uci_target.shape

uci_target['target'] = uci_target[0]

uci_target = uci_target.drop([0], axis = 1)

# Concatenating on the columns axis; the two sets need to be concatenated before sampling so that targets
# match row entries.
uci_madelon = pd.concat([uci_train, uci_target], axis = 1) 


uci_madelon.shape # checking the correct shape of the data; data is ready to be sampled from

### Set 1

uci_madelon1 = uci_madelon.sample(n=200)

uci_madelon1.shape # checking the correct shape, 200 rows and 501 columns.

### Set 2

uci_madelon2 = uci_madelon.sample(n=200)

uci_madelon2.shape

### Set 3

uci_madelon3 = uci_madelon.sample(n=200)

uci_madelon3.shape

### Pickling Data Sets

joblib.dump(uci_madelon1, 'uci_madelon1new.pkl') # 10% sample

joblib.dump(uci_madelon2, 'uci_madelon2new.pkl') # 10% sample

joblib.dump(uci_madelon3, 'uci_madelon3new.pkl') # 10% sample

joblib.dump(uci_madelon, 'uci_madelonALL.pkl') # the whole data set

## Josh's Madelon Server

from ConnectionHelper import ConnectionHelperUtil

### Set 1

josh_conn = ConnectionHelperUtil.ConnectionHelper('34.211.227.227', 'postgres', 'postgres')

j_madelon1 = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon1a = j_madelon1 # fetching four(4) 5000 random sample sets

j_madelon1b = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon1c = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon1d = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon1 = pd.concat([j_madelon1a, 
                        j_madelon1b, 
                        j_madelon1c, 
                        j_madelon1d]) # Concatenating all four to make the first set of 20,000 random samples
                                      # Will repeat this process for the other two sets

j_madelon1.shape # checking shape to confirm the data set has 20,000 rows

### Set 2

j_madelon2a = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon2b = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon2c = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon2d = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon2 = pd.concat([j_madelon2a,
                        j_madelon2b,
                        j_madelon2c,
                        j_madelon2d])

j_madelon2.shape

### Set 3

j_madelon3a = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon3b = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon3c = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon4d = josh_conn.connect_and_fetch("SELECT * FROM madelon ORDER BY random() LIMIT 5000", df=True)

j_madelon3d = j_madelon4d

j_madelon3 = pd.concat([j_madelon3a,
                        j_madelon3b,
                        j_madelon3c,
                        j_madelon3d])

j_madelon3.shape

### Pickling Data Sets

joblib.dump(j_madelon1a, 'j_madelon1a.pkl')

joblib.dump(j_madelon1, 'j_madelon1.pkl') # pickling all three data sets to be able to access them in other notebooks

joblib.dump(j_madelon2, 'j_madelon2.pkl')

joblib.dump(j_madelon3, 'j_madelon3.pkl')

small change
