==> Notebook_1_EDA.ipynb <==
# Notebook 1 - EDA

%matplotlib inline

from sklearn.externals import joblib
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

uci_madelon = joblib.load("./pickles/uci_madelonALL.pkl")

j_madelon = joblib.load('./pickles/j_madelon_test.pkl')

## UCI

uci_madelon.head()

uci_madelon.shape

## Josh's

j_madelon.head()

j_madelon.shape

small change
