==> Notebook_3_Feature_Selection.ipynb <==
# Notebook 3 - Feature Selection

!pip install tqdm

from sklearn.pipeline import make_pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.neighbors import KNeighborsClassifier, KNeighborsRegressor
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.svm import LinearSVC
from sklearn.feature_selection import SelectKBest, SelectFromModel, RFE, chi2, f_regression

import numpy as np
import pandas as pd
import tqdm

import pickle
from sklearn.externals import joblib

# Feature Selection

## on UCI

uci_madelon = joblib.load('./pickles/uci_madelon1new.pkl')

uci_madelon_no_target = uci_madelon.drop(['target'], axis = 1)

uci_madelon.head()

### Correlation Matrix

corr_matrix = uci_madelon_no_target.corr()

n = corr_matrix.shape[0]

for i in range(n):
    corr_matrix.iloc[i,i] = 0

corr_matrix = corr_matrix.abs()

max_corr = corr_matrix.max()

feats = list(corr_matrix.loc[:, max_corr > .6].columns)
print(len(feats))
feats

joblib.dump(feats, 'uci_feats.pkl')

### SelectKBest

skb = SelectKBest()

 def find_KBest(data):
    '''
    Find the 20 best features from a set of data. Data input should come without a target column.
    '''
    for col in data.columns:
        
        X = data.drop([col], axis = 1)
        y = data[col]
        
        X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.33)
        
        fit = SelectKBest(score_func= f_regression, k=20).fit(X_train,y_train)
    
        skb_feats = list(np.where(fit.get_support())[0])
        
    return skb_feats

skb_feats = find_KBest(uci_madelon_no_target)
skb_feats

### SelectFromModel

X = uci_madelon_no_target
y = uci_madelon['target']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.33)

sfm_logr = SelectFromModel(LogisticRegression(), threshold = '2.85*mean')

sfm_logr.fit(X_train, y_train)

sfm_logr_feats = np.where(sfm_logr.get_support())[0]

sfm_logr_feats

### RFE

rfe = RFE(LogisticRegression(), n_features_to_select=20, step = 5, verbose=1)

rfe.fit(X_train, y_train)

rfe_feats = np.where(rfe.get_support())[0]
rfe_feats

### Mean R2

def calculate_r_2_for_feature(data, feature, regression_method):
    '''
    Given a dataset, a feature, and a regression method, return the R2 for that feature.
    '''
    new_data = data.drop(feature, axis=1)
    X_train, X_test, y_train, y_test = train_test_split(new_data, data[feature], test_size=0.33)

    regressor = regression_method
    regressor.fit(X_train, y_train)

    score = regressor.score(X_test, y_test)
    return score


def mean_r2_for_feature(data, feature, regression_method):
    '''
    Take 3 iterations of caclulating the R2 for a feature, then take the average.
    '''
    scores = []
    
    for _ in range(3):
        scores.append(calculate_r_2_for_feature(data, feature, regression_method))   
    
    scores = np.array(scores)
    mean_scores = scores.mean()
    
    return mean_scores 

def return_feature_scores(data, regression_method):
    '''
    Returns the scores of all features. 
    '''
    feature_scores = []
    columns = data.columns
    
    for col in columns:
        mean_r2 = mean_r2_for_feature(data, col, regression_method)
        
        if mean_r2 > 0:
            feature_scores.append((col, mean_r2))
            
        else:
            pass
        
    return feature_scores

return_feature_scores(uci_madelon, KNeighborsRegressor(n_jobs = -1))

## on Josh's

`j_madelon_test.pkl` is a sample set with shape `(2000, 1002)`

j_madelon = joblib.load('./pickles/j_madelon_test.pkl')
j_madelon = j_madelon.drop(['_id'], axis = 1)

Xj = j_madelon.drop(['target'], axis = 1)
yj = j_madelon['target']

### Correlation Matrix

corr_matrix_j = j_madelon.corr()

np.fill_diagonal(corr_matrix_j.values, 0)

corr_matrix_j = corr_matrix_j.abs()

max_corr_j = corr_matrix_j.max()

feats_j = list(corr_matrix_j.loc[:, max_corr_j > .6].columns)
print(len(feats_j))
feats_j

[257, 269, 308, 315, 336, 341, 395, 504, 526, 639, 681, 701, 724, 736, 769, 808, 829, 867, 920, 956]

joblib.dump(feats_j, 'josh_feats.pkl')

return_feature_scores(j_madelon, Log(n_jobs = -1))

## Dimension Reduction

### UCI

Decided to go with the 20 features found from the correlation matrix technique.

`28, 48, 64, 105, 128, 153, 241, 281, 318, 336, 338, 378, 433, 442, 451, 453, 455, 472, 475, 493`


from sklearn.decomposition import PCA

pca_uci = PCA(n_components=5)

pca_uci.fit(uci_madelon[feats])

pd.DataFrame(pca_uci.components_)

pca_uci.explained_variance_

uci_madelon_pca = pca_uci.transform(uci_madelon[feats])

uci_madelon_pca_df = pd.DataFrame(uci_madelon_pca)

uci_madelon_pca_df.head()

### Josh's

pca_josh = PCA(n_components=5)

pca_josh.fit(j_madelon[feats_j])

pd.DataFrame(pca_josh.components_)

pca_josh.explained_variance_

josh_madelon_pca = pca_josh.transform(j_madelon[feats_j])

josh_madelon_pca_df = pd.DataFrame(josh_madelon_pca)

josh_madelon_pca_df.head()

small change
