==> Notebook_4_Build_Model.ipynb <==
# Notebook 4 - Building/GridSearching Models

from sklearn.pipeline import make_pipeline, Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.neighbors import KNeighborsClassifier
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.svm import LinearSVC, SVC
from sklearn.feature_selection import SelectKBest, SelectFromModel, RFE, chi2, f_regression
from sklearn.decomposition import PCA

import numpy as np
import pandas as pd
import tqdm

import pickle
from sklearn.externals import joblib

## Fitting Models on UCI

uci_madelon = joblib.load('./pickles/uci_madelonALL.pkl')

feats = joblib.load('./pickles/uci_feats.pkl')

uci_mad_feats = uci_madelon[feats]

y = uci_madelon['target']
X = uci_mad_feats

pipe_tree = Pipeline([('scaler', StandardScaler(with_mean=False)),
                      ('pca', PCA(n_components=5)),
                      ('estimator', DecisionTreeClassifier())])

pipe_logr = Pipeline([('scaler', StandardScaler(with_mean=False)),
                      ('pca', PCA(n_components=5)),
                      ('estimator', LogisticRegression())])

pipe_knc  = Pipeline([('scaler', StandardScaler(with_mean=False)),
                      ('pca', PCA(n_components=5)),
                      ('estimator', KNeighborsClassifier())])

pipe_svc  = Pipeline([('scaler', StandardScaler(with_mean=False)),
                      ('pca', PCA(n_components=5)),
                      ('estimator', SVC())])

pipe_rfc  = Pipeline([('scaler', StandardScaler(with_mean=False)),
                      ('pca', PCA(n_components=5)),
                      ('estimator', RandomForestClassifier())])

tree_params = {'estimator__max_depth' : [2,3,4,5,6,7,8,9,10],
               'estimator__max_features' : ['sqrt', 'log2', 'auto'],}

logr_params = {'estimator__C' : np.logspace(-3,3,7),
               'estimator__penalty' : ['l1', 'l2'],}

knc_params  = {'estimator__n_neighbors' : [3,5,11,],}

svc_params  = {'estimator__C' : np.logspace(-3,3,4),
               'estimator__kernel' : ['rbf']}

rfc_params  = {'estimator__n_estimators' : [25,50,75,100],
               'estimator__max_features' : ['sqrt', 'auto', 'log2'],}

grid_tree = GridSearchCV(pipe_tree, tree_params, verbose = 10)
grid_logr = GridSearchCV(pipe_logr, logr_params, verbose = 10)
grid_knc  = GridSearchCV(pipe_knc, knc_params, verbose = 10)
grid_svc  = GridSearchCV(pipe_svc, svc_params, verbose = 10)
grid_rfc  = GridSearchCV(pipe_rfc, rfc_params, verbose = 10)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.33)

uci_grid_tree = grid_tree.fit(X_train, y_train)
uci_grid_logr = grid_logr.fit(X_train, y_train)
uci_grid_knc  = grid_knc.fit(X_train, y_train)
uci_grid_svc  = grid_svc.fit(X_train, y_train)
uci_grid_rfc  = grid_rfc.fit(X_train, y_train)

uci_grid_tree 
uci_grid_logr 
uci_grid_knc  
uci_grid_svc  
uci_grid_rfc  

grid_logr_train_score_uci = uci_grid_logr.score(X_train, y_train)
grid_logr_test_score_uci  = uci_grid_logr.score(X_test, y_test)

grid_tree_train_score_uci = uci_grid_tree.score(X_train, y_train)
grid_tree_test_score_uci  = uci_grid_tree.score(X_test, y_test)

grid_knc_train_score_uci  = uci_grid_knc.score(X_train, y_train)
grid_knc_test_score_uci   = uci_grid_knc.score(X_test, y_test)

grid_svc_train_score_uci  = uci_grid_svc.score(X_train, y_train)
grid_svc_test_score_uci   = uci_grid_svc.score(X_test, y_test)

grid_rfc_train_score_uci  = uci_grid_rfc.score(X_train, y_train)
grid_rfc_test_score_uci   = uci_grid_rfc.score(X_test, y_test)

uci_grid_results = []

uci_grid_results.append({'model': 'LogisticRegressionGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_logr_train_score_uci})

uci_grid_results.append({'model': 'LogisticRegressionGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_logr_test_score_uci})

uci_grid_results.append({'model': 'DecisionTreeClassifierGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_tree_train_score_uci})

uci_grid_results.append({'model': 'DecisionTreeClassifierGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_tree_test_score_uci})

uci_grid_results.append({'model': 'KNeighborsClassifierGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_knc_train_score_uci})

uci_grid_results.append({'model': 'KNeighborsClassifierGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_knc_test_score_uci})

uci_grid_results.append({'model': 'SVCGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_svc_train_score_uci})

uci_grid_results.append({'model': 'SVCGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_svc_test_score_uci})

uci_grid_results.append({'model': 'RandomForestClassifierGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_rfc_train_score_uci})

uci_grid_results.append({'model': 'RandomForestClassifierGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_rfc_test_score_uci})

uci_grid_results_df = pd.DataFrame(uci_grid_results)
uci_grid_results_df

## Fitting Models on Josh's

j_madelon = joblib.load('./pickles/j_madelon1.pkl')

j_madelon = j_madelon.drop(['_id'], axis = 1)

j_feats = joblib.load('./pickles/josh_feats.pkl')

j_mad_feats = j_madelon[j_feats]

Xj = j_mad_feats
yj = j_madelon['target']

X_train_j, X_test_j, y_train_j, y_test_j = train_test_split(Xj, yj, test_size = 0.33)

j_grid_tree = grid_tree.fit(X_train_j, y_train_j)
j_grid_tree

j_grid_logr = grid_logr.fit(X_train_j, y_train_j)
j_grid_logr

j_grid_knc = grid_knc.fit(X_train_j, y_train_j)
j_grid_knc

j_grid_svc = grid_svc.fit(X_train_j, y_train_j)
j_grid_svc

j_grid_rfc = grid_rfc.fit(X_train_j, y_train_j)
j_grid_rfc

grid_logr_train_score_j = grid_logr.score(X_train_j, y_train_j)
grid_logr_test_score_j  = grid_logr.score(X_test_j, y_test_j)

grid_tree_train_score_j = grid_tree.score(X_train_j, y_train_j)
grid_tree_test_score_j  = grid_tree.score(X_test_j, y_test_j)

grid_knc_train_score_j  = grid_knc.score(X_train_j, y_train_j)
grid_knc_test_score_j   = grid_knc.score(X_test_j, y_test_j)

grid_svc_train_score_j  = grid_svc.score(X_train_j, y_train_j)
grid_svc_test_score_j   = grid_svc.score(X_test_j, y_test_j)

grid_rfc_train_score_j  = grid_rfc.score(X_train_j, y_train_j)
grid_rfc_test_score_j   = grid_rfc.score(X_test_j, y_test_j)

josh_grid_results = []

josh_grid_results.append({'model': 'LogisticRegressionGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_logr_train_score_j})

josh_grid_results.append({'model': 'LogisticRegressionGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_logr_test_score_j})

josh_grid_results.append({'model': 'DecisionTreeClassifierGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_tree_train_score_j})

josh_grid_results.append({'model': 'DecisionTreeClassifierGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_tree_test_score_j})

josh_grid_results.append({'model': 'KNeighborsClassifierGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_knc_train_score_j})

josh_grid_results.append({'model': 'KNeighborsClassifierGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_knc_test_score_j})

josh_grid_results.append({'model': 'SVCGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_svc_train_score_j})

josh_grid_results.append({'model': 'SVCGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_svc_test_score_j})

josh_grid_results.append({'model': 'RandomForestClassifierGrid',
                         'dataset' : 'train',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_rfc_train_score_j})

josh_grid_results.append({'model': 'RandomForestClassifierGrid',
                         'dataset' : 'test',
                         'preprocessing': 'scaled, PCA',
                         'score': grid_rfc_test_score_j})

josh_grid_results_df = pd.DataFrame(josh_grid_results)
josh_grid_results_df

small change
